refactor: rename src package to quanttradeai
