+++ Makefile
@@ -1,16 +1,17 @@
 .PHONY: format format-check lint test pipeline
 
 format:
-	poetry run black src/
+	poetry run black quanttradeai/
 
 format-check:
-	poetry run black --check src/
+	poetry run black --check quanttradeai/
 
 lint:
-	poetry run flake8 src/
+	poetry run flake8 quanttradeai/
 
 test:
 	poetry run pytest
 
 pipeline:
 	poetry run quanttradeai train -c config/model_config.yaml
+
