+++ docs/api/backtesting.md
@@ -18,7 +18,7 @@ Simulates trades using label signals.
 
 **Example:**
 ```python
-from src.backtest.backtester import simulate_trades
+from quanttradeai.backtest.backtester import simulate_trades
 
 # Simulate trades with stop-loss and take-profit
 df_with_trades = simulate_trades(df, stop_loss_pct=0.02, take_profit_pct=0.04)
@@ -40,7 +40,7 @@ Calculates basic performance metrics for a strategy.
 
 **Example:**
 ```python
-from src.backtest.backtester import compute_metrics
+from quanttradeai.backtest.backtester import compute_metrics
 
 # Calculate performance metrics
 metrics = compute_metrics(df_with_trades, risk_free_rate=0.02)
@@ -53,7 +53,7 @@ print(f"Max Drawdown: {metrics['max_drawdown']:.4f}")
 
 ### Basic Backtesting
 ```python
-from src.backtest.backtester import simulate_trades, compute_metrics
+from quanttradeai.backtest.backtester import simulate_trades, compute_metrics
 
 # Simulate trades
 df_trades = simulate_trades(df_labeled)
@@ -68,8 +68,8 @@ print(f"Max Drawdown: {metrics['max_drawdown']:.2%}")
 
 ### Backtesting with Risk Management
 ```python
-from src.backtest.backtester import simulate_trades, compute_metrics
-from src.trading.risk import apply_stop_loss_take_profit
+from quanttradeai.backtest.backtester import simulate_trades, compute_metrics
+from quanttradeai.trading.risk import apply_stop_loss_take_profit
 
 # Apply risk management rules
 df_with_risk = apply_stop_loss_take_profit(
@@ -91,7 +91,7 @@ print(f"Max Drawdown: {metrics['max_drawdown']:.2%}")
 
 ### Multi-Asset Backtesting
 ```python
-from src.backtest.backtester import simulate_trades, compute_metrics
+from quanttradeai.backtest.backtester import simulate_trades, compute_metrics
 
 # Backtest multiple assets
 results = {}
@@ -166,7 +166,7 @@ print(f"Average Loss: {losing_trades['strategy_return'].mean():.2%}")
 
 ### Stop-Loss and Take-Profit
 ```python
-from src.trading.risk import apply_stop_loss_take_profit
+from quanttradeai.trading.risk import apply_stop_loss_take_profit
 
 # Apply different risk management scenarios
 scenarios = [
