+++ docs/api/trading.md
@@ -18,7 +18,7 @@ Applies stop-loss and take-profit rules to trading signals.
 
 **Example:**
 ```python
-from src.trading.risk import apply_stop_loss_take_profit
+from quanttradeai.trading.risk import apply_stop_loss_take_profit
 
 # Apply 2% stop-loss and 4% take-profit
 df_with_risk = apply_stop_loss_take_profit(df, stop_loss_pct=0.02, take_profit_pct=0.04)
@@ -45,7 +45,7 @@ Calculates position size based on account risk parameters.
 
 **Example:**
 ```python
-from src.trading.risk import position_size
+from quanttradeai.trading.risk import position_size
 
 # Calculate position size
 qty = position_size(capital=10000, risk_per_trade=0.02, stop_loss_pct=0.05, price=150.0)
@@ -56,7 +56,7 @@ print(f"Position size: {qty} shares")
 
 ### Basic Risk Management
 ```python
-from src.trading.risk import apply_stop_loss_take_profit
+from quanttradeai.trading.risk import apply_stop_loss_take_profit
 
 # Apply risk management to trading signals
 df_with_risk = apply_stop_loss_take_profit(
@@ -72,7 +72,7 @@ print(f"Risk-adjusted signals: {risk_adjusted_signals}")
 
 ### Dynamic Position Sizing
 ```python
-from src.trading.risk import position_size
+from quanttradeai.trading.risk import position_size
 
 # Calculate position sizes for different scenarios
 scenarios = [
@@ -88,7 +88,7 @@ for capital, risk, sl, price in scenarios:
 
 ### Portfolio Risk Management
 ```python
-from src.trading.risk import apply_stop_loss_take_profit, position_size
+from quanttradeai.trading.risk import apply_stop_loss_take_profit, position_size
 
 # Manage risk across multiple positions
 portfolio = {
