+++ tests/models/test_classifier.py
@@ -2,13 +2,8 @@ import unittest
 import pandas as pd
 import numpy as np
 import os
-import sys
 
-sys.path.insert(
-    0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "..", "src"))
-)
-
-from models.classifier import MomentumClassifier  # noqa: E402
+from quanttradeai.models.classifier import MomentumClassifier
 
 
 class TestMomentumClassifier(unittest.TestCase):
